app_search_tool/requirements.txt
apispec==6.3.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cors==1.0.1
coverage==7.1.0
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.0
Faker==16.8.1
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-openapi3==2.5.3
flask-rest-api==0.17.1
Flask-SQLAlchemy==3.0.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
mqtt==0.0.1
packaging==23.0
paho-mqtt==1.6.1
password-strength==0.0.3.post2
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.13
PyJWT==2.0.0
PySocks==1.7.1
pytest==7.2.1
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dateutil==2.8.2
requests==2.28.2
requests-file==1.5.1
schedule==1.2.0
six==1.16.0
SQLAlchemy==2.0.0
tldextract==3.4.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
webargs==8.3.0
Werkzeug==2.2.2
zope.event==4.6
zope.interface==5.5.2
